Move HDF5 dataset outside gymnos library
