+++ gymnos/datasets/__init__.py
@@ -5,6 +5,6 @@ from .mte import MTE
 from .dogs_vs_cats import DogsVsCats
 from .data_usage_test import DataUsageTest
 from .unusual_data_usage_test import UnusualDataUsageTest
-from .dataset import Dataset, HDF5Dataset
+from .dataset import Dataset, DatasetInfo, Array, ClassLabel
 from .synthetic_digits import SyntheticDigits
 from .rock_paper_scissors import RockPaperScissors
